test: generate stubs and move local tests
