+++ test_data/compiled-workflows/local_execution_pipeline_with_ignore_upstream_failure.yaml
