+++ test_data/compiled-workflows/local_execution_pipeline_with_k8s_only_methods.yaml
