+++ test_data/compiled-workflows/local_execution_pipeline_with_oneof.yaml
