+++ test_data/compiled-workflows/pipeline_with_volume.yaml
@@ -7,13 +7,13 @@ spec:
     parameters:
     - name: kubernetes-comp-consumer
       value: '{"pvcMount":[{"mountPath":"/data","pvcNameParameter":{"taskOutputParameter":{"outputParameterKey":"name","producerTask":"createpvc"}},"taskOutputParameter":{"outputParameterKey":"name","producerTask":"createpvc"}}]}'
-    - name: components-4ca7976f7f3d605d80b9d8bd28e9ae3e49e2b0263940665ccd83fdada3282e8c
+    - name: components-7c670709fc771e1d79ebc2c954ea9e008eb5216e226dfa1594e8dec49167ac06
       value: '{"executorLabel":"exec-consumer","outputDefinitions":{"parameters":{"Output":{"parameterType":"STRING"}}}}'
-    - name: implementations-4ca7976f7f3d605d80b9d8bd28e9ae3e49e2b0263940665ccd83fdada3282e8c
+    - name: implementations-7c670709fc771e1d79ebc2c954ea9e008eb5216e226dfa1594e8dec49167ac06
       value: '{"args":["--executor_input","{{$}}","--function_to_execute","consumer"],"command":["sh","-c","\nif
         ! [ -x \"$(command -v pip)\" ]; then\n    python3 -m ensurepip || python3
         -m ensurepip --user || apt-get install python3-pip\nfi\n\nPIP_DISABLE_PIP_VERSION_CHECK=1
-        python3 -m pip install --quiet --no-warn-script-location ''kfp==2.14.3'' ''--no-deps''
+        python3 -m pip install --quiet --no-warn-script-location ''kfp==2.15.2'' ''--no-deps''
         ''typing-extensions\u003e=3.7.4,\u003c5; python_version\u003c\"3.9\"'' \u0026\u0026
         \"$0\" \"$@\"\n","sh","-ec","program_path=$(mktemp -d)\n\nprintf \"%s\" \"$0\"
         \u003e \"$program_path/ephemeral_component.py\"\n_KFP_RUNTIME=true python3
@@ -22,8 +22,10 @@ spec:
         consumer() -\u003e str:\n    with open(''/data/file.txt'', ''r'') as file:\n        content
         = file.read()\n    print(content)\n    return content\n\n"],"image":"registry.access.redhat.com/ubi9/python-311:latest"}'
     - name: components-98f254581598234b59377784d6cbf209de79e0bcda8013fe4c4397b5d3a26767
-      value: '{"executorLabel":"exec-createpvc","inputDefinitions":{"parameters":{"access_modes":{"description":"Values
-        for ``PersistentVolumeClaim.spec.accessModes``.","parameterType":"LIST"},"annotations":{"description":"Annotations
+      value: '{"executorLabel":"exec-createpvc","inputDefinitions":{"parameters":{"access_modes":{"description":"AccessModes
+        to request for the provisioned PVC. May\nbe one or more of ``''ReadWriteOnce''``,
+        ``''ReadOnlyMany''``,\n``''ReadWriteMany''``, or ``''ReadWriteOncePod''``.
+        Corresponds to\n`PersistentVolumeClaim.spec.accessModes\n\u003chttps://kubernetes.io/docs/concepts/storage/persistent-volumes/#access-modes\u003e`_.","parameterType":"LIST"},"annotations":{"description":"Annotations
         for the PVC''s metadata. Corresponds to\n`PersistentVolumeClaim.metadata.annotations\n\u003chttps://kubernetes.io/docs/reference/kubernetes-api/config-and-storage-resources/persistent-volume-claim-v1/#PersistentVolumeClaim\u003e`_.","isOptional":true,"parameterType":"STRUCT"},"data_source":{"description":"Specifies
         the data source for the PVC. This can be used to\ncreate a PVC from a VolumeSnapshot
         or to clone an existing PVC.\nShould be a dictionary with ``''apiGroup''``,
@@ -51,13 +53,13 @@ spec:
       value: '{"image":"argostub/deletepvc"}'
     - name: kubernetes-comp-producer
       value: '{"pvcMount":[{"mountPath":"/data","pvcNameParameter":{"taskOutputParameter":{"outputParameterKey":"name","producerTask":"createpvc"}},"taskOutputParameter":{"outputParameterKey":"name","producerTask":"createpvc"}}]}'
-    - name: components-71bc0395c46e8bddfaad94bfe774d00d048562e7c07f1b9eab9ec6d3655cc5d2
+    - name: components-7cefc145506c478b829283f76980bbba7892efa3ed3b73e4dfe0d6be991d14a8
       value: '{"executorLabel":"exec-producer","outputDefinitions":{"parameters":{"Output":{"parameterType":"STRING"}}}}'
-    - name: implementations-71bc0395c46e8bddfaad94bfe774d00d048562e7c07f1b9eab9ec6d3655cc5d2
+    - name: implementations-7cefc145506c478b829283f76980bbba7892efa3ed3b73e4dfe0d6be991d14a8
       value: '{"args":["--executor_input","{{$}}","--function_to_execute","producer"],"command":["sh","-c","\nif
         ! [ -x \"$(command -v pip)\" ]; then\n    python3 -m ensurepip || python3
         -m ensurepip --user || apt-get install python3-pip\nfi\n\nPIP_DISABLE_PIP_VERSION_CHECK=1
-        python3 -m pip install --quiet --no-warn-script-location ''kfp==2.14.3'' ''--no-deps''
+        python3 -m pip install --quiet --no-warn-script-location ''kfp==2.15.2'' ''--no-deps''
         ''typing-extensions\u003e=3.7.4,\u003c5; python_version\u003c\"3.9\"'' \u0026\u0026
         \"$0\" \"$@\"\n","sh","-ec","program_path=$(mktemp -d)\n\nprintf \"%s\" \"$0\"
         \u003e \"$program_path/ephemeral_component.py\"\n_KFP_RUNTIME=true python3
@@ -299,11 +301,11 @@ spec:
       - arguments:
           parameters:
           - name: component
-            value: '{{workflow.parameters.components-4ca7976f7f3d605d80b9d8bd28e9ae3e49e2b0263940665ccd83fdada3282e8c}}'
+            value: '{{workflow.parameters.components-7c670709fc771e1d79ebc2c954ea9e008eb5216e226dfa1594e8dec49167ac06}}'
           - name: task
             value: '{"cachingOptions":{"enableCache":true},"componentRef":{"name":"comp-consumer"},"dependentTasks":["createpvc","producer"],"taskInfo":{"name":"consumer"}}'
           - name: container
-            value: '{{workflow.parameters.implementations-4ca7976f7f3d605d80b9d8bd28e9ae3e49e2b0263940665ccd83fdada3282e8c}}'
+            value: '{{workflow.parameters.implementations-7c670709fc771e1d79ebc2c954ea9e008eb5216e226dfa1594e8dec49167ac06}}'
           - name: task-name
             value: consumer
           - name: parent-dag-id
@@ -355,11 +357,11 @@ spec:
       - arguments:
           parameters:
           - name: component
-            value: '{{workflow.parameters.components-71bc0395c46e8bddfaad94bfe774d00d048562e7c07f1b9eab9ec6d3655cc5d2}}'
+            value: '{{workflow.parameters.components-7cefc145506c478b829283f76980bbba7892efa3ed3b73e4dfe0d6be991d14a8}}'
           - name: task
             value: '{"cachingOptions":{"enableCache":true},"componentRef":{"name":"comp-producer"},"dependentTasks":["createpvc"],"taskInfo":{"name":"producer"}}'
           - name: container
-            value: '{{workflow.parameters.implementations-71bc0395c46e8bddfaad94bfe774d00d048562e7c07f1b9eab9ec6d3655cc5d2}}'
+            value: '{{workflow.parameters.implementations-7cefc145506c478b829283f76980bbba7892efa3ed3b73e4dfe0d6be991d14a8}}'
           - name: task-name
             value: producer
           - name: parent-dag-id
