+++ test_data/sdk_compiled_pipelines/valid/local_execution/local_execution_pipeline_with_oneof.py
