+++ test_data/sdk_compiled_pipelines/valid/local_execution/pipeline_with_caching.yaml
